% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{limpiecito}
\alias{limpiecito}
\title{Limpieza de caracteres o strings.}
\usage{
limpiecito(x)
}
\arguments{
\item{x}{un elemento o vector character.}
}
\description{
Es una función para limpiar espacios por demás al inicio, final o intermedio. Asimismo
transforma las letras minúsculas en mayúsculas para un mejor tratamiento.
}
\details{
Esta función se usará para la limpieza de base de datos.
}
\examples{
limpiecito(c("    raul", " yo uso    R"))
# "RAUL" "YO USO R"
}
